Coaxis.Accounts.Resources.UserInterest (coaxis v0.1.0)

Functions

can_create(actor, params_or_opts \\ %{}, opts \\ [])

can_create?(actor, params_or_opts \\ %{}, opts \\ [])

can_destroy(actor, record, params_or_opts \\ %{}, opts \\ [])

can_destroy?(actor, record, params_or_opts \\ %{}, opts \\ [])

can_get_by_id(actor, id, params_or_opts \\ %{}, opts \\ [])

can_get_by_id?(actor, id, params_or_opts \\ %{}, opts \\ [])

can_read(actor, params_or_opts \\ %{}, opts \\ [])

can_read?(actor, params_or_opts \\ %{}, opts \\ [])

changeset_to_create(params_or_opts \\ %{}, opts \\ [])

changeset_to_destroy(record, params_or_opts \\ %{}, opts \\ [])

create(params_or_opts \\ %{}, opts \\ [])

Calls the create action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:changeset (term/0) - A changeset to seed the action with.

create!(params_or_opts \\ %{}, opts \\ [])

Calls the create action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:changeset (term/0) - A changeset to seed the action with.

default_short_name()

destroy(record, params_or_opts \\ %{}, opts \\ [])

Calls the destroy action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

destroy!(record, params_or_opts \\ %{}, opts \\ [])

Calls the destroy action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

entities(arg1)

fetch_opt(arg1, arg2)

get_by_id(id, params_or_opts \\ %{}, opts \\ [])

Calls the by_id action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:query (term/0) - A query to seed the action with.

:load (term/0) - Adds a load statement to the query before passing it to the action.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

get_by_id!(id, params_or_opts \\ %{}, opts \\ [])

Calls the by_id action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:query (term/0) - A query to seed the action with.

:load (term/0) - Adds a load statement to the query before passing it to the action.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

input(opts)

@spec input(values :: map() | Keyword.t()) :: map() | no_return()

Validates that the keys in the provided input are valid for at least one action on the resource.

Raises a KeyError error at compile time if not. This exists because generally a struct should only ever be created by Ash as a result of a successful action. You should not be creating records manually in code, e.g %MyResource{value: 1, value: 2}. Generally that is fine, but often with embedded resources it is nice to be able to validate the keys that are being provided, e.g

Resource
|> Ash.Changeset.for_create(:create, %{embedded: EmbeddedResource.input(foo: 1, bar: 2)})
|> MyApp.Api.create()

input(opts, action)

@spec input(values :: map() | Keyword.t(), action :: atom()) :: map() | no_return()

Same as input/1, except restricts the keys to values accepted by the action provided.

persisted()

persisted(arg1)

persisted(arg1, default)

primary_key_matches?(left, right)

query_to_get_by_id(id, params_or_opts \\ %{}, opts \\ [])

query_to_read(params_or_opts \\ %{}, opts \\ [])

read(params_or_opts \\ %{}, opts \\ [])

Calls the read action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:query (term/0) - A query to seed the action with.

:load (term/0) - Adds a load statement to the query before passing it to the action.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

read!(params_or_opts \\ %{}, opts \\ [])

Calls the read action on the Coaxis.Accounts.Resources.UserInterest resource.

Options

:tenant (term/0) - Set the tenant of the query/changeset

:context (term/0) - Set context on the query/changeset

:actor (term/0) - set the actor for authorization

:tracer (term/0) - set the tracer for the action

:authorize? (boolean/0) - Whether or not to perform authorization. The default behavior depends on the api configuration.

:verbose? (boolean/0) - a flag to toggle verbose output from the internal Ash engine (for debugging)

:timeout (timeout/0) - A timeout to apply to the operation.

:query (term/0) - A query to seed the action with.

:load (term/0) - Adds a load statement to the query before passing it to the action.

:not_found_error? (boolean/0) - Whether or not to return or raise a NotFound error or to return nil when a get? action/interface is called.

spark_dsl_config()

spark_is()